requirements.txt
# Bindings for the hermes protocol 
hermes-python>=0.1
requests
serial
RPi.GPIO
# More dependency goes here.. 
